Move scaffold logic to its own directory
